requirements.txt
numpy == 1.19.5
opencv-contrib-python == 4.5.3.56
opencv-python == 4.5.3.56
Pillow == 8.4.0
matplotlib = 3.4.3
